fix: fixing post and creating scrapping

+++ src/repository/publicationRepository.js
@@ -1,9 +1,8 @@
-import db from "../database/db.js";
+import db from '../database/db.js';
+import urlMetadata from 'url-metadata';
 
 export async function findUser(token) {
-  const result = await db.query(`SELECT * FROM sessions WHERE token = $1`, [
-    token,
-  ]);
+  const result = await db.query(`SELECT * FROM sessions WHERE token = $1`, [token]);
   return result;
 }
 
@@ -30,31 +29,36 @@ export async function insertHashtags(insertIndex, hashtags) {
 
 export async function getPublications() {
   const result = await db.query(`
-    SELECT json_build_object(
-      'id', posts.id,
-      'userId', users.id,
-      'name', users.name,
-      'avatarImage', users.avatar_url,
-      'descriptionPost', posts.description,
-      'linkPost', posts.link,
-      'qtyLikesPost', COUNT(likes.id),
-      'idUsersLike', array_agg(json_build_object(
-      'id', likes.user_id
-      )),
-      'hashtags', array_agg(json_build_object(
-          'id', hashtags.id,
-          'nameHashtag', hashtags.name
-      ))
-    )
-    FROM users
-    JOIN posts
-    ON users.id = posts.user_id
-    LEFT JOIN hashtags
-    ON posts.id = hashtags.post_id
-    LEFT JOIN likes
-    ON posts.id = likes.post_id
-    GROUP BY users.id, posts.id
-    ORDER BY posts.id DESC;`);
+  SELECT json_build_object(
+    'id', posts.id,
+    'userId', users.id,
+    'name', users.name,
+    'avatarImage', users.avatar_url,
+    'descriptionPost', posts.description,
+    'linkPost', posts.link,
+    'titleLinkPost', info_link_post.title_link,
+    'descriptionLinkPost', info_link_post.description_link,
+    'imageLinkPost', info_link_post.image_link,
+    'qtyLikesPost', COUNT(likes.id),
+    'idUsersLike', array_agg(json_build_object(
+    'id', likes.user_id
+    )),
+    'hashtags', array_agg(json_build_object(
+        'id', hashtags.id,
+        'nameHashtag', hashtags.name
+    ))
+  )
+  FROM users
+  JOIN posts
+  ON users.id = posts.user_id
+  LEFT JOIN hashtags
+  ON posts.id = hashtags.post_id
+  LEFT JOIN likes
+  ON posts.id = likes.post_id
+  LEFT JOIN info_link_post
+  ON posts.id = info_link_post.post_id
+  GROUP BY users.id, posts.id, info_link_post.id
+  ORDER BY posts.id DESC;`);
 
   return result;
 }
@@ -74,22 +78,19 @@ export async function queryVerifyUserId(postId, userId) {
 }
 
 export async function deleteHashtagByIdPost(postId) {
-  return await db.query("DELETE FROM hashtags WHERE post_id = $1", [postId]);
+  return await db.query('DELETE FROM hashtags WHERE post_id = $1', [postId]);
 }
 
 export async function deletePostById(postId) {
-  return await db.query("DELETE FROM posts WHERE id = $1", [postId]);
+  return await db.query('DELETE FROM posts WHERE id = $1', [postId]);
 }
 
 export async function updatePostByid(description, postId) {
-  return await db.query("UPDATE posts SET description=$1 WHERE id = $2;", [
-    description,
-    postId,
-  ]);
+  return await db.query('UPDATE posts SET description=$1 WHERE id = $2;', [description, postId]);
 }
 
 export async function findUsersLikByPostId(id, limit) {
-  return db.query(
+  return await db.query(
     `
   SELECT posts.id AS "postId",
       posts.user_id AS "createdUserId",
@@ -108,15 +109,16 @@ export async function findUsersLikByPostId(id, limit) {
 }
 
 export async function insertLikeInPost(postId, userId) {
-  return db.query("INSERT INTO likes (post_id,user_id) VALUES ($1, $2);", [
-    postId,
-    userId,
-  ]);
+  return await db.query('INSERT INTO likes (post_id,user_id) VALUES ($1, $2);', [postId, userId]);
 }
 
 export async function deleteLikeInPost(postId, userId) {
-  return db.query("DELETE FROM likes WHERE post_id = $1 and user_id = $2;", [
-    postId,
-    userId,
-  ]);
+  return await db.query('DELETE FROM likes WHERE post_id = $1 and user_id = $2;', [postId, userId]);
+}
+
+export async function insertInfoLinkScrapping(post_id, title_link, description_link, image_link) {
+  return await db.query(
+    `INSERT INTO  info_link_post (post_id, title_link, description_link,image_link) VALUES ($1, $2, $3, $4)`,
+    [post_id, title_link, description_link, image_link]
+  );
 }
